rewrite pngConverter
